+++ sdk/mojoauth-api.js
@@ -5,28 +5,20 @@
 module.exports = function (config) {
 
   function signinWithMagicLink(email) {
-    var queryParameters = {};
-
-    queryParameters.api_Key = config.apiKey;
-
     var bodyParameters = {};
     bodyParameters.email = email;
 
     var resourcePath = 'users/magiclink';
 
-    return config.request('POST', resourcePath, queryParameters, bodyParameters);
+    return config.request('POST', resourcePath, "", bodyParameters);
   }
 
   function verifyToken(token) {
-    var queryParameters = {};
-
-    queryParameters.api_Key = config.apiKey;
-
     var resourcePath = 'token/jwks';
 
     return new Promise(function (resolve, reject) {
 
-      var jwksPromise =  config.request('GET', resourcePath, queryParameters, "");
+      var jwksPromise =  config.request('GET', resourcePath, "", "");
 
       jwksPromise.then(function (jwksResponse) {
         const jwktopem = require('jwk-to-pem');
@@ -53,29 +45,23 @@ module.exports = function (config) {
 
     queryParameters.api_Key = config.apiKey;
     queryParameters.state_id = stateId;
-
+    
     var resourcePath = 'users/status';
 
     return config.request('GET', resourcePath, queryParameters, "");
   }
 
   function signinWithEmailOTP(email) {
-    var queryParameters = {};
-
-    queryParameters.api_Key = config.apiKey;
-
     var bodyParameters = {};
     bodyParameters.email = email;
 
     var resourcePath = 'users/emailotp';
 
-    return config.request('POST', resourcePath, queryParameters, bodyParameters);
+    return config.request('POST', resourcePath, "", bodyParameters);
   }
 
   function verifyOtp(otp) {
     var queryParameters = {};
-
-    queryParameters.api_Key = config.apiKey;
     queryParameters.otp = otp;
 
     var resourcePath = 'users/emailotp/verify';