+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Version 1.2.0
 
+- Enabled use of MojoAuth API secret.
 - Renamed `mojoAPI.pingStaus` to `pingStatus`, retaining a deprecated `pingStaus` alias to prevent breaking changes.
 - Fixed `mojoAPI.pingStatus` to use the correct `state_id` query parameter name.
+
 
 ## Version 1.0.0
 Released on **March 23, 2021**
  - Added MojoAuth APIs
-<br>
+
 See the documentation [here](https://mojoauth.com/docs/)